// src/plugins/setubal/geoanalitycs-plugin/gql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Upload: { input: any; output: any; }
};
// ...
export enum FileCacheType {
  JSON = 'JSON',
  TEXT = 'TEXT'
}
// ...
export type GeoActivityData = {
  __typename?: 'GeoActivityData';
  averageDuration?: Maybe<Scalars['Int']['output']>;
  days?: Maybe<Array<Scalars['Int']['output']>>;
  total?: Maybe<Scalars['Int']['output']>;
  visits?: Maybe<Array<Scalars['Int']['output']>>;
};

export type GeoDataInput = {
  month: Scalars['Int']['input'];
  origin: Scalars['String']['input'];
  period: Scalars['String']['input'];
  refresh?: InputMaybe<Scalars['Boolean']['input']>;
  year: Scalars['Int']['input'];
};

export type GeoLabelsAndData = {
  __typename?: 'GeoLabelsAndData';
  data?: Maybe<Array<Scalars['Int']['output']>>;
  labels?: Maybe<Array<Scalars['String']['output']>>;
};

export type GeoMetadata = {
  __typename?: 'GeoMetadata';
  end_date?: Maybe<Scalars['String']['output']>;
  start_date?: Maybe<Scalars['String']['output']>;
};

export type GeoStatsData = {
  __typename?: 'GeoStatsData';
  busiest_day?: Maybe<Scalars['String']['output']>;
  busiest_hour?: Maybe<Scalars['String']['output']>;
  metadata?: Maybe<GeoMetadata>;
  regulars?: Maybe<Scalars['Int']['output']>;
  residents?: Maybe<Scalars['Int']['output']>;
  tourists?: Maybe<GeoVisitors>;
  visitors?: Maybe<GeoVisitors>;
  visits_day?: Maybe<Scalars['Int']['output']>;
  visits_hour?: Maybe<Scalars['Int']['output']>;
};

export type GeoVisitors = {
  __typename?: 'GeoVisitors';
  internationals?: Maybe<Scalars['Int']['output']>;
  nationals?: Maybe<Scalars['Int']['output']>;
  total?: Maybe<Scalars['Int']['output']>;
};

export type GeoVisitsPerHour = {
  __typename?: 'GeoVisitsPerHour';
  hours?: Maybe<Array<Scalars['String']['output']>>;
  visits?: Maybe<Array<Scalars['Int']['output']>>;
};